backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytesseract
pdf2image
Pillow
agno
weasyprint
openpyxl
pytest
httpx
python-dotenv
